Feature/shortform 영상 삭제 로직 추가

+++ src/main/java/com/mingles/metamingle/shortform/command/application/service/ShortFormFirebaseService.java
@@ -1,7 +1,9 @@
 package com.mingles.metamingle.shortform.command.application.service;
 
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.storage.*;
+import com.google.cloud.storage.Blob;
+import com.google.cloud.storage.BlobId;
+import com.google.cloud.storage.Bucket;
+import com.google.cloud.storage.Storage;
 import com.google.firebase.cloud.StorageClient;
 import com.mingles.metamingle.shortform.command.domain.aggregate.vo.MemberNoVO;
 import com.mingles.metamingle.shortform.command.application.dto.response.CreateShortFormResponse;
@@ -9,18 +11,20 @@ import com.mingles.metamingle.shortform.command.application.dto.response.DeleteS
 import com.mingles.metamingle.shortform.command.domain.aggregate.entity.ShortForm;
 import com.mingles.metamingle.shortform.command.domain.repository.ShortFormCommandRepository;
 import com.mingles.metamingle.shortform.command.domain.service.ShortFormCommandDomainService;
+import com.mingles.metamingle.shortform.command.infrastructure.service.ApiInteractiveMovieCommandService;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.io.IOUtils;
 import org.jcodec.api.FrameGrab;
 import org.jcodec.api.JCodecException;
 import org.jcodec.common.io.NIOUtils;
 import org.jcodec.common.model.Picture;
 import org.jcodec.scale.AWTUtil;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StreamUtils;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
@@ -28,6 +32,7 @@ import java.io.*;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 import java.util.UUID;
 
 
@@ -46,19 +51,21 @@ public class ShortFormFirebaseService {
 
     private final ShortFormCommandDomainService shortFormCommandDomainService;
 
+    private final ApiInteractiveMovieCommandService apiInteractiveMovieCommandService;
+
     // 숏폼 생성
     @Transactional
     public CreateShortFormResponse createShortForm(MultipartFile file, String title, String description) throws IOException, JCodecException {
         String fileKeyName = createFileName(file.getOriginalFilename()); // 파일 이름을 고유한 파일 이름으로 교체
 
-        GoogleCredentials credentials = GoogleCredentials.fromStream(new FileInputStream(firebaseConfigPath));
-        Storage storage = StorageOptions.newBuilder().setCredentials(credentials).build().getService();
-        BlobId blobId = BlobId.of(bucketName, fileKeyName);
-        BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType("video/mp4").build();
-
+        Bucket bucket = StorageClient.getInstance().bucket(bucketName);
         InputStream inputStream = file.getInputStream();
-        byte[] bytes = IOUtils.toByteArray(inputStream);
-        Blob blob = storage.create(blobInfo, bytes);
+        Blob blob = bucket.create(fileKeyName, inputStream, file.getContentType());
+        Blob getBlob = bucket.get(fileKeyName);
+
+        System.out.println("blob shortForm = " + getBlob.getMediaLink());
+
+
         inputStream.close();
 
         String url = bucketUrl + fileKeyName + "?alt=media";
@@ -86,26 +93,53 @@ public class ShortFormFirebaseService {
         ShortForm shortForm = shortFormCommandRepository.findById(shortFormNo)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지않는 숏폼입니다."));
 
+        List<Long> deletedInteractiveMovieNos = null;
+
+        if (shortForm.getIsInteractive()) {
+            deletedInteractiveMovieNos = apiInteractiveMovieCommandService.deleteInteractiveMovieWithShortFormNo(shortFormNo);
+        }
+
         // 사용자 확인 로직 (본인 확인)
+
+        // firebase storage의 영상, 썸네일 삭제
+        String thumbnailName = shortForm.getThumbnailUrl()
+                                        .replace(bucketUrl, "")
+                                        .replace("%2F", "/")
+                                        .replace("?alt=media", "");
+        String videoName = shortForm.getUrl()
+                                    .replace(bucketUrl, "")
+                                    .replace("?alt=media", "");
+
+        BlobId blobIdThumbnail = BlobId.of(bucketName, thumbnailName);
+        BlobId blobIdVideo = BlobId.of(bucketName, videoName);
+
+        Storage storage = StorageClient.getInstance().bucket(bucketName).getStorage();
+        Bucket bucket = StorageClient.getInstance().bucket(bucketName);
+
+        if (!storage.delete(blobIdThumbnail)) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "숏폼 썸네일 삭제 실패 ");
+        }
+
+        if (!storage.delete(blobIdVideo)) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "숏폼 무비 영상 삭제 실패 ");
+        }
 
         shortFormCommandRepository.delete(shortForm);
 
-        return new DeleteShortFormResponse(shortFormNo);
+        return new DeleteShortFormResponse(shortFormNo, deletedInteractiveMovieNos);
     }
 
     // 인터랙티브 무비와 관련된 숏폼 생성
     @Transactional
-    public ShortForm createShortFormWithInterativeMovie(MultipartFile file, String title, String description) throws IOException, JCodecException {
+    public ShortForm createShortFormWithInteractiveMovie(MultipartFile file, String title, String description) throws IOException, JCodecException {
         String fileKeyName = createFileName(file.getOriginalFilename()); // 파일 이름을 고유한 파일 이름으로 교체
 
-        GoogleCredentials credentials = GoogleCredentials.fromStream(new FileInputStream(firebaseConfigPath));
-        Storage storage = StorageOptions.newBuilder().setCredentials(credentials).build().getService();
-        BlobId blobId = BlobId.of(bucketName, fileKeyName);
-        BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType("video/mp4").build();
-
+        Bucket bucket = StorageClient.getInstance().bucket(bucketName);
         InputStream inputStream = file.getInputStream();
-        byte[] bytes = IOUtils.toByteArray(inputStream);
-        Blob blob = storage.create(blobInfo, bytes);
+        Blob blob = bucket.create(fileKeyName, inputStream, file.getContentType());
+
+        System.out.println("blob interactiveMovie = " + blob.getMediaLink());
+
         inputStream.close();
 
         String url = bucketUrl + fileKeyName + "?alt=media";
@@ -160,12 +194,11 @@ public class ShortFormFirebaseService {
         // 이미지 포맷을 "jpeg"로 변경
         ImageIO.write(bufferedImage, "jpeg", baos);
         byte[] thumbnailBytes = baos.toByteArray();
-        Storage storage = StorageClient.getInstance().bucket(bucketName).getStorage();
+        Bucket bucket = StorageClient.getInstance().bucket(bucketName);
         String thumbnailKey = "thumbnails/" + fileKeyName.replace(".mp4", ".jpeg");
+        Blob blob = bucket.create(thumbnailKey, thumbnailBytes, "image/jpeg");
 
-        BlobId blobId = BlobId.of(bucketName, thumbnailKey);
-        BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType("image/jpeg").build();
-        Blob thumbnailBlob = storage.create(blobInfo, thumbnailBytes);
+        System.out.println("blob thumbnail = " + blob.getMediaLink());
 
         return bucketUrl + thumbnailKey.replace("/", "%2F") + "?alt=media";
     }
